tests/requirements-test.txt
# Core test framework
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27

# Database clients (for integration tests)
motor>=3.3
qdrant-client>=1.7
redis>=5.0
minio>=7.2

# LLM evaluation
ragas>=0.1
langsmith>=0.1

# Load testing
locust>=2.20

# JWT generation for test auth
python-jose[cryptography]>=3.3

# Environment variables
python-dotenv>=1.0

# Sentence embeddings (for answer consistency tests)
sentence-transformers>=2.2

# PDF generation (for test fixtures)
fpdf2>=2.7
